+++ core/src/components/modal/modal-interface.ts
@@ -1,5 +1,3 @@
-import type { JSXBase } from '@stencil/core/internal';
-
 import type { AnimationBuilder, ComponentProps, ComponentRef, FrameworkDelegate, Mode } from '../../interface';
 
 export interface ModalOptions<T extends ComponentRef = ComponentRef> {
@@ -11,7 +9,12 @@ export interface ModalOptions<T extends ComponentRef = ComponentRef> {
   cssClass?: string | string[];
   delegate?: FrameworkDelegate;
   animated?: boolean;
+  /**
+ * If `true`, the modal can be swiped to dismiss. Only applies in iOS mode.
+ * @deprecated - To prevent modals from dismissing, use canDismiss instead.
+ */
   swipeToClose?: boolean;
+  canDismiss?: boolean | (() => Promise<boolean>);
 
   mode?: Mode;
   keyboardClose?: boolean;
@@ -33,4 +36,15 @@ export interface ModalAnimationOptions {
   backdropBreakpoint?: number;
 }
 
-export type ModalAttributes = JSXBase.HTMLAttributes<HTMLElement>;
+export interface ModalBreakpointChangeEventDetail {
+  breakpoint: number;
+}
+
+export interface ModalCustomEvent extends CustomEvent {
+  target: HTMLIonModalElement;
+}
+
+/**
+ * @deprecated - Use { [key: string]: any } directly instead.
+ */
+export type ModalAttributes = { [key: string]: any };