+++ core/src/components/modal/test/basic/modal.e2e.ts
@@ -1,0 +1,182 @@
+import { expect } from '@playwright/test';
+import { test } from '@utils/test/playwright';
+
+test.describe('modal: focus trapping', () => {
+  test('focus should be trapped inside of modal', async ({ page, browserName }) => {
+    await page.goto('/src/components/modal/test/basic');
+    const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+
+    await page.click('#basic-modal');
+
+    await ionModalDidPresent.next();
+
+    /**
+     * The default WebKit behavior is to
+     * highlight items on webpages with Option-Tab.
+     * See "Press Tab to highlight each item on a webpage"
+     * in Safari Preferences > Advanced.
+     */
+    if (browserName === 'webkit') {
+      await page.keyboard.down('Alt');
+    }
+    await page.keyboard.press('Tab');
+
+    const dismissButton = await page.locator('ion-button.dismiss');
+    expect(dismissButton).toBeFocused();
+
+    await page.keyboard.down('Shift');
+    await page.keyboard.press('Tab');
+    await page.keyboard.up('Shift');
+
+    expect(dismissButton).toBeFocused();
+    await page.keyboard.press('Tab');
+
+    if (browserName === 'webkit') {
+      await page.keyboard.up('Alt');
+    }
+
+    expect(dismissButton).toBeFocused();
+  });
+
+  test('focus should be returned to previously focused element when dismissing modal', async ({
+    page,
+    browserName,
+  }) => {
+    await page.goto('/src/components/modal/test/basic');
+    const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+    const ionModalDidDismiss = await page.spyOnEvent('ionModalDidDismiss');
+    const modalButton = await page.locator('#basic-modal');
+
+    if (browserName === 'webkit') {
+      await page.keyboard.down('Alt');
+    }
+
+    // Focus #basic-modal button
+    await page.keyboard.press('Tab');
+    expect(modalButton).toBeFocused();
+
+    if (browserName === 'webkit') {
+      await page.keyboard.up('Alt');
+    }
+
+    await page.keyboard.press('Space');
+    await ionModalDidPresent.next();
+
+    await page.keyboard.press('Escape');
+    await ionModalDidDismiss.next();
+
+    expect(modalButton).toBeFocused();
+  });
+});
+
+test.describe('modal: rendering', () => {
+  test('should not have visual regressions', async ({ page }) => {
+    await page.goto('/src/components/modal/test/basic');
+
+    const ionModalWillDismiss = await page.spyOnEvent('ionModalWillDismiss');
+    const ionModalDidDismiss = await page.spyOnEvent('ionModalDidDismiss');
+    const ionModalWillPresent = await page.spyOnEvent('ionModalWillPresent');
+    const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+
+    await page.click('#basic-modal');
+
+    await ionModalWillPresent.next();
+    await ionModalDidPresent.next();
+
+    const modal = await page.locator('ion-modal');
+
+    await page.setIonViewport();
+
+    expect(await page.screenshot()).toMatchSnapshot(`modal-basic-present-${page.getSnapshotSettings()}.png`);
+
+    await modal.evaluate((el: HTMLIonModalElement) => {
+      el.dismiss();
+    });
+
+    await ionModalWillDismiss.next();
+    await ionModalDidDismiss.next();
+
+    expect(await page.screenshot()).toMatchSnapshot(`modal-basic-dismiss-${page.getSnapshotSettings()}.png`);
+  });
+});
+
+test.describe('modal: htmlAttributes inheritance', () => {
+  test('should correctly inherit attributes on host', async ({ page }) => {
+    await page.goto('/src/components/modal/test/basic');
+    const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+
+    await page.click('#basic-modal');
+
+    await ionModalDidPresent.next();
+
+    const modal = await page.locator('ion-modal');
+
+    const attribute = await modal.getAttribute('data-testid');
+    expect(attribute).toBe('basic-modal');
+  });
+});
+
+test.describe('modal: backdrop', () => {
+  test.beforeEach(async ({ page }) => {
+    await page.goto('/src/components/modal/test/basic');
+  });
+
+  test('it should dismiss the modal when clicking the backdrop', async ({ page }, testInfo) => {
+    test.skip(
+      testInfo.project.name === 'Mobile Chrome' || testInfo.project.name === 'Mobile Safari',
+      'The modal covers the entire screen on mobile, so the backdrop is not clickable.'
+    );
+
+    const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+    const ionModalDidDismiss = await page.spyOnEvent('ionModalDidDismiss');
+
+    await page.click('#basic-modal');
+    await ionModalDidPresent.next();
+
+    await page.mouse.click(20, 20);
+    await ionModalDidDismiss.next();
+  });
+});
+
+test.describe('modal: incorrect usage', () => {
+  test.beforeEach(async ({ page }) => {
+    await page.goto('/src/components/modal/test/basic');
+  });
+
+  test('it should warn when setting a breakpoint on a non-sheet modal', async ({ page }) => {
+    const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+
+    const warnings: string[] = [];
+
+    page.on('console', (ev) => {
+      if (ev.type() === 'warning') {
+        warnings.push(ev.text());
+      }
+    });
+
+    await page.click('#basic-modal');
+    await ionModalDidPresent.next();
+
+    const modal = await page.locator('ion-modal');
+    await modal.evaluate((el: HTMLIonModalElement) => {
+      el.setCurrentBreakpoint(0.5);
+    });
+
+    expect(warnings.length).toBe(1);
+    expect(warnings[0]).toBe('[Ionic Warning]: setCurrentBreakpoint is only supported on sheet modals.');
+  });
+
+  test('it should return undefined when getting the breakpoint on a non-sheet modal', async ({ page }) => {
+    const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+
+    await page.click('#basic-modal');
+    await ionModalDidPresent.next();
+
+    const modal = await page.locator('ion-modal');
+    const breakpoint = await modal.evaluate((el: HTMLIonModalElement) => {
+      return el.getCurrentBreakpoint();
+    });
+
+    expect(breakpoint).toBe(undefined);
+  });
+});