+++ core/src/interface.d.ts
@@ -12,6 +12,7 @@ export * from './components/checkbox/checkbox-interface';
 export * from './components/datetime/datetime-interface';
 export * from './components/infinite-scroll/infinite-scroll-interface';
 export * from './components/input/input-interface';
+export * from './components/item/item-interface';
 export * from './components/item-sliding/item-sliding-interface';
 export * from './components/loading/loading-interface';
 export * from './components/menu/menu-interface';