+++ core/src/utils/input-shims/hacks/hide-caret.ts
@@ -5,7 +5,7 @@ import { isFocused, relocateInput } from './common';
 export const enableHideCaretOnScroll = (
   componentEl: HTMLElement,
   inputEl: HTMLInputElement | HTMLTextAreaElement | undefined,
-  scrollEl: HTMLIonContentElement | undefined
+  scrollEl: HTMLElement | undefined
 ) => {
   if (!scrollEl || !inputEl) {
     return () => {