+++ core/stencil.config.ts
@@ -166,10 +166,6 @@ export const config: Config = {
       includeGlobalScripts: false
     },
     {
-      type: 'docs-readme',
-      strict: true
-    },
-    {
       type: 'docs-json',
       file: '../docs/core.json'
     },
@@ -263,7 +259,9 @@ export const config: Config = {
     pixelmatchThreshold: 0.05,
     waitBeforeScreenshot: 20,
     moduleNameMapper: {
-      "@utils/test": ["<rootDir>/src/utils/test/utils"]
+      "@utils/test": ["<rootDir>/src/utils/test/utils"],
+      "@utils/logging": ["<rootDir>/src/utils/logging"],
+
     },
     emulate: [
       {