+++ packages/vue/test-app/tests/e2e/specs/nested.js
@@ -4,10 +4,6 @@ describe('Nested', () => {
     cy.ionPageVisible('nestedchild');
   });
 
-  it('should show first page', () => {
-    cy.ionPageVisible('nestedchild');
-  });
-
   it('should go to second page', () => {
     cy.get('#nested-child-two').click();
     cy.ionPageVisible('nestedchildtwo');
@@ -31,4 +27,26 @@ describe('Nested', () => {
     cy.ionPageDoesNotExist('tab1');
     cy.ionPageVisible('home');
   });
+})
+
+describe('Nested - Replace', () => {
+  it('should replace a route but still be able to go back to main outlet', () => {
+    cy.visit('http://localhost:8080');
+
+    cy.routerPush('/nested');
+    cy.ionPageHidden('home');
+    cy.ionPageVisible('nestedchild');
+
+    cy.routerReplace('/nested/two');
+    cy.ionPageDoesNotExist('nestedchild');
+    cy.ionPageVisible('nestedchildtwo');
+
+    /**
+     * ionBackClick does not handle nested pages
+     * with multiple back buttons
+     */
+    cy.get('#routeroutlet-back-button').click();
+    cy.ionPageDoesNotExist('nestedchildtwo');
+    cy.ionPageVisible('home');
+  })
 })